+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Upd: Update colored `TextStyle` code writer
 * Upd: Update token code writers
+* Upd: Update build context extension code writer
 
 ## 0.2.2